cleanup / refactor for clarity
