+++ src/atoms/MapAtoms.tsx
@@ -6,7 +6,7 @@ import { Theme } from '@/theme/Theme';
 
 const MAPBOX_TOKEN = process.env.REACT_APP_MAPBOX_ENV;
 
-export const Map = ({ startingCoordinates, children, mapStyle }: MapProps) => {
+export const Map = ({ startingCoordinates, children, mapStyle, interactive = true }: MapProps) => {
   const [viewport, setViewport] = React.useState({
     latitude: startingCoordinates?.lat ?? 42.286,
     longitude: startingCoordinates?.lng ?? -71.088,
@@ -22,10 +22,10 @@ export const Map = ({ startingCoordinates, children, mapStyle }: MapProps) => {
 
   return (
     <ReactMapGL
-      { ...viewport }
-      { ...settings }
-      mapboxApiAccessToken={ MAPBOX_TOKEN }
-      onViewportChange={ (viewport: any) => setViewport(viewport) }
+      {...viewport}
+      {...settings}
+      mapboxApiAccessToken={MAPBOX_TOKEN}
+      onViewportChange={ (viewport: any) => interactive && setViewport(viewport)}
       mapStyle={ mapStyle ?? Theme.map.light }
     >
       { children }
@@ -80,5 +80,22 @@ export const MapGeoJsonSource = ({ data, id, type, color, visible }: Omit<MapGeo
     <Source type='geojson' data={ data }>
       <Layer { ...layer }/>
     </Source>
+  );
+};
+
+export function Pin({ size = 20, color = Theme.colors.orange}: {size: number, color: string}) {
+  const ICON = 'M8 16s6-5.686 6-10A6 6 0 0 0 2 6c0 4.314 6 10 6 10zm0-7a3 3 0 1 1 0-6 3 3 0 0 1 0 6z';
+  
+  const pinStyle = {
+    // cursor: 'pointer',
+    fill: color,
+    stroke: 'none',
+    display: 'block',
+  };
+
+  return (
+    <svg height={size} viewBox="0 0 16 16" style={pinStyle}>
+      <path d={ICON} />
+    </svg>
   );
 };
