+++ src/atoms/MapStoryAtoms.tsx
@@ -1,24 +1,5 @@
 import React from 'react';
 import styled from '@emotion/styled';
-import { Theme } from '@/theme/Theme';
-
-
-export function Pin({ size = 20, color = Theme.colors.orange}: {size: number, color: string}) {
-  const ICON = 'M8 16s6-5.686 6-10A6 6 0 0 0 2 6c0 4.314 6 10 6 10zm0-7a3 3 0 1 1 0-6 3 3 0 0 1 0 6z';
-  
-  const pinStyle = {
-    // cursor: 'pointer',
-    fill: color,
-    stroke: 'none',
-    display: 'block',
-  };
-
-  return (
-    <svg height={size} viewBox="0 0 16 16" style={pinStyle}>
-      <path d={ICON} />
-    </svg>
-  );
-};
 
 export function Chapter({ id, content, title, image, isActive, hidden, alignment }: ChapterProps) {
   return (
